refactor: move the create database process to its intended database manager
